+++ src/main/java/com/blz/bookstore/repository/BookStoreRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import com.blz.bookstore.model.BookListDataModel;
 
 @Repository
-public interface BookStoreRepository extends JpaRepository<BookListDataModel, Integer> {
+public interface BookStoreRepository extends JpaRepository<BookListDataModel, Long> {
 }
